modify files and code structure about central-arm-func
